[shell][refactor] structure
