+++ TestShell/TestShell/TestShell.vcxproj
@@ -130,15 +130,20 @@
     <None Include="packages.config" />
   </ItemGroup>
   <ItemGroup>
+    <ClCompile Include="common.cpp" />
     <ClCompile Include="executor.h" />
     <ClCompile Include="main.cpp" />
+    <ClCompile Include="shellCmdError.cpp" />
+    <ClCompile Include="shellCmdFlush.cpp" />
+    <ClCompile Include="shellCmdHelp.cpp" />
+    <ClCompile Include="shellCmdInvalid.cpp" />
     <ClCompile Include="shellCmdParser.cpp" />
     <ClCompile Include="ssdExecutor.cpp" />
     <ClCompile Include="mockExecutor.cpp" />
-    <ClCompile Include="testShellCmdComplex.cpp" />
+    <ClCompile Include="shellCmdComplex.cpp" />
     <ClCompile Include="testShellCmdInterface.h" />
-    <ClCompile Include="testShellCmdRead.cpp" />
-    <ClCompile Include="testShellCmdWrite.cpp" />
+    <ClCompile Include="shellCmdRead.cpp" />
+    <ClCompile Include="shellCmdWrite.cpp" />
     <ClCompile Include="test_complex.cpp" />
     <ClCompile Include="test_ShellCmdParser.cpp" />
     <ClCompile Include="test_read.cpp">
@@ -155,9 +160,11 @@
     </ClCompile>
   </ItemGroup>
   <ItemGroup>
+    <ClInclude Include="common.h" />
     <ClInclude Include="shellCmdParser.h" />
     <ClInclude Include="shellFixture.cpp" />
-    <ClInclude Include="testShell.cpp" />
+    <ClInclude Include="shell.cpp" />
+    <ClInclude Include="testShellCmd.h" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
   <ImportGroup Label="ExtensionTargets">
