+++ TestShell/TestShell/main.cpp
@@ -1,5 +1,5 @@
 #include "gmock/gmock.h"
-#include "testShell.cpp"
+#include "shell.cpp"
 #include "ssdExecutor.cpp"
 
 #ifdef _DEBUG
