+++ TestShell/TestShell/mockExecutor.cpp
@@ -1,11 +1,12 @@
 #pragma once
-#include "gmock/gmock.h"
-#include "executor.h"
 #include <string>
+
+#include "executor.h"
+#include "gmock/gmock.h"
 
 using namespace testing;
 
-class MockSSD :public Executor {
+class MockSSD : public Executor {
 public:
 	MOCK_METHOD(std::string, readFromSSD, (int), (override));
 	MOCK_METHOD(std::string, writeToSSD, (int, int), (override));
