+++ TestShell/TestShell/shellCmdError.cpp
@@ -1,0 +1,9 @@
+#include "testShellCmd.h"
+#include "common.h"
+
+TestShellErrorCmd::TestShellErrorCmd() {
+}
+
+void TestShellErrorCmd::run() {
+	std::cout << "unknown command \n";
+}
