+++ TestShell/TestShell/shellFixture.cpp
@@ -1,7 +1,7 @@
 #pragma once
 #include "gmock/gmock.h"
 #include "mockExecutor.cpp"
-#include "testShell.cpp"
+#include "shell.cpp"
 
 using namespace testing;
 class ShellFixture : public Test {
